backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
websockets
python-multipart
pydantic
google-generativeai
openpyxl
python-dotenv
